requirement.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
python-jose[openssl] 